requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
anthropic>=0.25.0
supabase>=2.4.0
python-dotenv>=1.0.0
rich>=13.7.0
schedule>=1.2.0